Tests are roughly twice faster with this option.

Before:
```
Duration  7.58s (transform 859ms, setup 441ms, collect 8.55s, tests 2.61s, environment 15.28s, prepare 3.19s)
```

After:
```
Duration  3.06s (transform 1.03s, setup 344ms, collect 7.36s, tests 2.08s, environment 1.58s, prepare 4.25s)
```