@cexbrayat 2023-11-08
CI builds are twice faster as well